refactor & rename : channelApis 내부 api파일들 channel/**.ts로 분리
