+++ src/apis/channel/getChannelList.ts
@@ -1,0 +1,14 @@
+import axiosInstance from '../api'
+
+import { GET_CHANNEL_LIST_PATH } from '@/utils/api_paths'
+
+async function getChannelList() {
+  try {
+    const request = await axiosInstance.get(GET_CHANNEL_LIST_PATH)
+    return request.data
+  } catch (error) {
+    throw new Error(`${error}`)
+  }
+}
+
+export default getChannelList
